Exclusive access to the response writer
This patch modifies an interface of the ResponseWriter. It removes
the method WriteHeader, in order to address the problem with Write
atomicity. As the ResponseWriter shares the same underlying connection
between multiple handlers, we un-intentionally put the Write-WriteHeader
method calls into the race condition. To mitigate this, I reduced
the count of methods, that provides ResponseWriter to one.