Integration test with a third-party switch (#51)
This patch introduces a concent of connection state (the same
as in the "net/http".Server type to handle the connection state
change. In a such way users will be able to catch the moment
of new switch connection.

Also within this patch was introduced a simple integration test,
that responds to OpenFlow switch hello and echo messages. It also
sends a multipart aggregation stats request.