Hello and Echo handlers
This patch introduces two help functions "EchoHandler" and
"HelloHandler" with a simple default implementation. Also to be able
write unit tests without the need to start a server this patch provides
"ofptest".ResposeRecorder type, that simply saves all requests into the
list, so they could be accessed validated in the unit tests.