Update documentation of port-related messages (#102)
This patch updates the documentation of the port description, port
statistics, port state messages.

It adds the missing kinds of the port configuration (those that
are relate to the STP).

It also updates the logic used to produce the human-redable
representation of the port configuration. Now, all options will be
consolidated into the space-separated list of strings.